setting respesctive folder for each middleware
